+++ src/irvine/oeis/a056/A056188.java
@@ -17,6 +17,6 @@ public class A056188 extends Sequence1 {
   @Override
   public Z next() {
     ++mN;
-    return Integers.SINGLETON.sum(1, mN, k -> Functions.GCD.l(mN, (long) k) == 1 ? Binomial.binomial(mN, k) : Z.ZERO);
+    return Integers.SINGLETON.sum(1, mN, k -> Functions.GCD.l(mN, k) == 1 ? Binomial.binomial(mN, k) : Z.ZERO);
   }
 }
