+++ src/irvine/oeis/a191/A191236.java
@@ -15,6 +15,6 @@ public class A191236 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A191236() {
-    super(0, n -> Integers.SINGLETON.sum(0, n, k -> Binomial.binomial(n, k).multiply(Functions.STIRLING2.z(2L * n - k, (long) n))));
+    super(0, n -> Integers.SINGLETON.sum(0, n, k -> Binomial.binomial(n, k).multiply(Functions.STIRLING2.z(2L * n - k, n))));
   }
 }
