+++ src/irvine/oeis/a345/A345344.java
@@ -23,6 +23,6 @@ public class A345344 extends AbstractSequence {
   @Override
   public Z next() {
     ++mN;
-    return Integers.SINGLETON.sumdiv(mN, d -> Predicates.SQUARE.is((long) d) ? Functions.BIG_OMEGA.z(mN / d) : Z.ZERO);
+    return Integers.SINGLETON.sumdiv(mN, d -> Predicates.SQUARE.is(d) ? Functions.BIG_OMEGA.z(mN / d) : Z.ZERO);
   }
 }
